Massive Optimisation of Custom Content
AKA: Past us massively overengineered the uninstall custom content handling and it was slow as all hell (affecting start up times too)

Testing with a Sonic '06 Randomiser Suite mod containing 6538 custom files, Version 3.42 would take two minutes to even start up if "Automatically Uninstall Mods" was active. A build with this change takes around 14 seconds to start up instead. As the same check is performed when installing mods with the option enabled, this also optimises the install time when a lot of custom files are present, as it no longer spends forever checking for them to remove first.